Tests for util.py
